+++ kong2kic/testdata/consumer-v3-output-expected.yaml
@@ -20,6 +20,7 @@ metadata:
   name: key-auth-example-user
 stringData:
   key: my_api_key
+type: Opaque
 ---
 apiVersion: v1
 kind: Secret
@@ -28,7 +29,7 @@ metadata:
     kubernetes.io/ingress.class: kong
   labels:
     konghq.com/credential: jwt
-  name: jwt-auth-example-user
+  name: jwt-example-user
 stringData:
   algorithm: HS256
   key: my_jwt_secret
@@ -43,6 +44,7 @@ stringData:
     ewIDAQAB
     -----END PUBLIC KEY-----
   secret: my_secret_key
+type: Opaque
 ---
 apiVersion: v1
 kind: Secret
@@ -55,6 +57,7 @@ metadata:
 stringData:
   password: my_basic_password
   username: my_basic_user
+type: Opaque
 ---
 apiVersion: v1
 kind: Secret
@@ -63,9 +66,10 @@ metadata:
     kubernetes.io/ingress.class: kong
   labels:
     konghq.com/credential: acl
-  name: acl-group-example-user
+  name: acl-example-user
 stringData:
   group: acl_group
+type: Opaque
 ---
 apiVersion: v1
 kind: Secret
@@ -83,9 +87,9 @@ type: Opaque
 apiVersion: configuration.konghq.com/v1
 credentials:
 - key-auth-example-user
-- jwt-auth-example-user
+- jwt-example-user
 - basic-auth-example-user
-- acl-group-example-user
+- acl-example-user
 - mtls-auth-example-user
 custom_id: "1234567890"
 kind: KongConsumer
